pyproject.toml
[project]
name = "llm-cognitive-bias-mitigation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
